requirements_local.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12    
cryptography==36.0.1
dj-database-url==0.5.0        
dj-static==0.0.6
dj-stripe==2.6.1
Django==4.0.3
django-admin-interface==0.19.0
django-colorfield==0.6.3      
django-cors-headers==3.11.0   
django-countries==7.3.2       
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-location-field==2.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
jsonfield==3.1.0
Pillow==9.0.1
pycodestyle==2.8.0
pycparser==2.21
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
stripe==2.67.0
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8